scripts/requirements.txt
# Python requirements for Firestore automation scripts

google-cloud-firestore>=2.0.0

